fix: increase e2e callback timeout and add per-role deadline
The manifest flow callback redirect timeout (10s) was too tight for
GitHub's app creation, which can be slow. Increased to 30s.

Also added a 2-minute context deadline per role in the app setup loop
so that failures produce a clear error instead of hanging until the
Go test timeout panics.

+++ e2e/admin/browser.go
@@ -178,8 +178,10 @@ func (b *PlaywrightBrowserOpener) handleCreateAppPage() error {
 	}
 
 	// Wait for redirect back to our callback URL.
+	// GitHub's manifest flow can be slow (app creation, key generation),
+	// so use a generous timeout.
 	if err := b.page.WaitForURL("**/callback**", playwright.PageWaitForURLOptions{
-		Timeout: playwright.Float(10000),
+		Timeout: playwright.Float(30000),
 	}); err != nil {
 		pageURL := b.page.URL()
 		if strings.Contains(pageURL, "/callback") || strings.Contains(pageURL, "127.0.0.1") {
